Move config package to internal
