+++ internal/config/testdata/deprecated/cache_memory_enabled.yml
