+++ internal/config/testdata/deprecated/cache_memory_items.yml
